requirements.txt
pip>=20.1b1
django
djangorestframework
requests==2.20.0